implement user roles, "user" and "admin", and add some admin features

+++ src/utils.js
@@ -1,3 +1,5 @@
+import { Task } from "./models/Task";
+
 export const getFromStorage = function (key) {
 	return JSON.parse(localStorage.getItem(key) || "[]");
 };
@@ -12,8 +14,10 @@ export const replaceStorage = function (key, newData) {
 	localStorage.setItem(key, JSON.stringify(newData));
 };
 
-export const generateTestUser = function (User) {
-	localStorage.clear();
-	const testUser = new User("test", "qwerty123");
-	User.save(testUser);
+export const generateFirstUsers = function (User) {
+	if (!getFromStorage(Task.storageKey).length) localStorage.clear();
+	const adminUser = new User("admin", "admin", "admin");
+	const mundaneUser = new User("test", "qwerty123", "user");
+	User.save(adminUser);
+	User.save(mundaneUser);
 };
